pyproject.toml
[project]
name = "reading-agent"
version = "1.0.0"
description = "Research article aggregation and summarization tool"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.32.0",
    "feedparser>=6.0.10",
    "beautifulsoup4>=4.12.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.1",
    "schedule>=1.2.0",
    "openai>=1.13.0",
    "httpx>=0.27.0",
    "tiktoken>=0.5.0",
    "lxml>=4.9.0",
    "email-validator>=2.1.0",
    # Async support
    "aiohttp>=3.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.1.0",
]

[project.scripts]
reading-agent = "src.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # function call in default argument
    "SIM108", # use ternary operator
]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
